u-boot: support booting fsbl for zc702
The upstream meta-xilinx does not include all necessary files to build the
first stage bootloader within the u-boot tree.  Add them here so the
boot.bin created by the normal build of core-image-* will allow booting of
the remainder of the system.

The following two files are required to build a bootable u-boot image and
come from the Xilinx embeddedsw repository:

   ps7_init.h: https://github.com/Xilinx/embeddedsw/blob/2c8f92039d3765a3b650286814b2612e8f1b099d/lib/sw_apps/zynq_fsbl/misc/zc702/ps7_init.h
   ps7_init.c: https://github.com/Xilinx/embeddedsw/blob/2c8f92039d3765a3b650286814b2612e8f1b099d/lib/sw_apps/zynq_fsbl/misc/zc702/ps7_init.c

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>