zynq-bsp: Disable Xilinx AFP DMA
The experimental Xilinx AFP driver has a bug causing it to hang when
booting in a single-processor configuration.  The cost is an extra
copy-to-user or copy-from-user on certain operations, so we'll disable the
driver until we have a resolution from the upstream provider.  This allows
us to support both SMP and AMP in the same configuration.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>