zc702: Update BSP configuration
uEnv.txt was out of date with renaming the image to zc702-zynq7-mel and
the default boot.bin was not being installed to the deploy directory
following the addition of mel-boot.bin to the layer.  Additionally,
mel-boot.bin was being copied from the incorrect directory after it'd been
moved.  Correct all three of these oversights.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>