zc702: Add minimal machine and dts
In order to support booting with the in-tree u-boot with a single kernel
configuration, we create a new MACHINE type with a dts that doesn't
describe any of the functionality that is disabled without the IP blob
from the first stage bootloader.

Configuring the zc702-zynq7-mel will have full support for the HDMI
display.  Configuring the zc702-zynq7-mel-minimal has support only for
serial console and headless configuration.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>