scripts: Add a script to automate preparing a bootable sd card
Simplify putting an image on a SD card for booting the ZC702 board with a
script that has reasonable defaults.

Note that this assumes you are building a non-HDMI-enabled configuration
(eg. MACHINE=zc702-zynq7-mel-minimal).  If you are building a HDMI-ready
version, you will need to pass a --devicetree option to point at
uImage-zynq-zc702-base-trd.dtb or something similar.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>