zc702: remove mel-boot.bin
The bitstream used to configure the FPGA to drive the HDMI display cannot
be used for more than an hour.  After that point, a reboot is necessary.
Customers wishing to use the HDMI display for full projects would need to
build a new bootloader using the Xilinx SDK and a different bitstream.
Therefore we are removing this demo bootloader.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>