config: default config should support MEL master
The defaults did not fully support MEL as a master configuration when
using MEMF.  Setup the default kernel configuration that will make it
possible to use MEL as a master out of the box.

Additionally, one ifdef was switching on the previous SMP flag rather than
using the proper OPENAMP configuration option.  Correct that at the same
time, since otherwise it could lead to an inconsistent configuration.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>