linux-adi: make B directory same as S to support CodeBench
Project configuration to debug kernel module in CodeBench
requires path to kernel source and vmlinux. And assumes that
certain file that is generated after build is located in source
folder, essentially requiring kernel build folder and source
folder to be the same. This commit makes B directory same as S
to facilitate CodeBench project configuration.

Jira-ID: SB-4606

Signed-off-by: Fahad Arslan <Fahad_Arslan@mentor.com>