linux-xlnx: build adi axi hdmi driver as built-in
When adi axi hdmi driver was being built as a module it was getting loaeded
at a fairly late stage during boot process and splash-screen was not coming
up. When we make it a part of kernel, fb device is created at a much earlier
stage.

Jira: http://jira.alm.mentorg.com:8080/browse/SB-6237

Signed-off-by: Fahad Usman <fahad_usman@mentor.com>