local.conf.append: remove perf in non GPLv3 build for aarch64 target
MPSoC build with GPLv3 in INCOMPATIBLE_LICENSE fails due to the version
of binutils selected as 2.17.50.0.12 which lacks support for AARCH64.
The AARCH64 support has been added in binutils version 2.23 onwards but
that is GPLv3 compatible. So we need to remove any dependencies on binutils
for non GPLv3 based build for aarch64 platform to succeed. Turns out that
perf depends on binutils and removing perf from
packagegroup-core-tools-profile for non GPLv3 compatible build results in
successful build.

JiraID: SB-8950

Signed-off-by: Yasir-Khan <yasir_khan@mentor.com>